refactor: Show Screen
